pyproject.toml
[tool.poetry]
name = "SUP"
version = "0.1.0"
description = ""
authors = ["Omelchenko Michael <socanime@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = {extras = ["standart"], version = "^0.23.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.1"
pydantic-settings = "^2.1.0"
argon2-cffi = "^23.1.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
json-log-formatter = "^0.5.2"
python-json-logger = "^2.0.7"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
ruff = "^0.1.14"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
